Move seq expr methods
